requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Reinforcement Learning Environment
gym[atari]>=0.26.0
autorom>=0.4.2,<0.5.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
imageio>=2.31.0

# Data Science and Visualization
numpy>=1.24.0
matplotlib>=3.7.0
tensorboard>=2.13.0

# Additional Utilities
tqdm>=4.65.0 